Support scale option
Often it is useful to run multiple containers to simulate an entier
fleet. With the volume based approach that doesn't work. We instead
mount the boot, data and state partitions as tmpfs.
This means that containers essentially become ephemeral, meaning the
balenaCloud device they map to is uniquely tied to the instance of the
container and stopping it leaves behind an orphaned device.

To work around that the tmpfs mounts can be replaced with a volume. And
the compose command has to be run multiple times (specifing different
project names).

Change-type: patch
Signed-off-by: Robert Günzler <robertg@balena.io>